bpf/tests/scapy: break up pkt_defs.py per test
This commit stubs the test-specific packet definitions into
dedicated `test_pkt_defs.py`, to have a cleaner packet definition
organization.

In order to break the circular dependency, the file structure is now:

* `pkt_defs_common.py`: common constants across tests
* `<test_name>_pkt_defs.py`: `test_name` specific packet/buffer
  definitions.
* `pkt_defs.py`: imports all packet/buffer definitions. This is still
  the file used by `parser.py` and `trace_diff_pkts.py`.
